+++ mobile/android/android-components/components/support/base/src/main/java/mozilla/components/support/base/ext/NotificationManagerCompat.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.ext
+package mozilla.components.support.base.ext
 
 import android.os.Build
 import androidx.core.app.NotificationChannelCompat