+++ mobile/android/android-components/components/support/base/src/test/java/mozilla/components/support/base/ext/NotificationManagerCompatTest.kt
@@ -2,51 +2,42 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.ext
+package mozilla.components.support.base.ext
 
 import android.os.Build
 import androidx.core.app.NotificationChannelCompat
 import androidx.core.app.NotificationManagerCompat
-import io.mockk.every
-import io.mockk.mockk
-import io.mockk.mockkStatic
-import org.junit.Assert.assertFalse
-import org.junit.Assert.assertTrue
-import org.junit.Before
+import junit.framework.TestCase.assertFalse
+import junit.framework.TestCase.assertTrue
+import mozilla.components.support.test.mock
+import mozilla.components.support.test.whenever
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mozilla.fenix.helpers.FenixRobolectricTestRunner
+import org.robolectric.RobolectricTestRunner
 import org.robolectric.util.ReflectionHelpers
 
-@RunWith(FenixRobolectricTestRunner::class)
+@RunWith(RobolectricTestRunner::class)
 class NotificationManagerCompatTest {
 
-    private lateinit var tested: NotificationManagerCompat
-
-    @Before
-    fun setup() {
-        tested = mockk(relaxed = true)
-
-        mockkStatic("org.mozilla.fenix.ext.NotificationManagerCompatKt")
-    }
+    private val tested: NotificationManagerCompat = mock()
 
     @Test
     fun `WHEN areNotificationsEnabled throws an exception THEN areNotificationsEnabledSafe returns false`() {
-        every { tested.areNotificationsEnabled() } throws RuntimeException()
+        whenever(tested.areNotificationsEnabled()).thenThrow(java.lang.RuntimeException())
 
         assertFalse(tested.areNotificationsEnabledSafe())
     }
 
     @Test
     fun `WHEN areNotificationsEnabled returns false THEN areNotificationsEnabledSafe returns false`() {
-        every { tested.areNotificationsEnabled() } returns false
+        whenever(tested.areNotificationsEnabled()).thenReturn(false)
 
         assertFalse(tested.areNotificationsEnabledSafe())
     }
 
     @Test
     fun `WHEN areNotificationsEnabled returns true THEN areNotificationsEnabledSafe returns true`() {
-        every { tested.areNotificationsEnabled() } returns true
+        whenever(tested.areNotificationsEnabled()).thenReturn(true)
 
         assertTrue(tested.areNotificationsEnabledSafe())
     }
@@ -60,8 +51,9 @@ class NotificationManagerCompatTest {
                 NotificationManagerCompat.IMPORTANCE_DEFAULT,
             ).build()
 
-        every { tested.getNotificationChannelCompat(testChannel) } returns notificationChannelCompat
-        every { tested.areNotificationsEnabled() } returns true
+        whenever(tested.areNotificationsEnabled()).thenReturn(true)
+        whenever(tested.getNotificationChannelCompat(testChannel))
+            .thenReturn(notificationChannelCompat)
 
         assertTrue(tested.isNotificationChannelEnabled(testChannel))
     }
@@ -75,8 +67,9 @@ class NotificationManagerCompatTest {
                 NotificationManagerCompat.IMPORTANCE_NONE,
             ).build()
 
-        every { tested.getNotificationChannelCompat(testChannel) } returns notificationChannelCompat
-        every { tested.areNotificationsEnabled() } returns true
+        whenever(tested.areNotificationsEnabled()).thenReturn(true)
+        whenever(tested.getNotificationChannelCompat(testChannel))
+            .thenReturn(notificationChannelCompat)
 
         assertFalse(tested.isNotificationChannelEnabled(testChannel))
     }
@@ -90,8 +83,9 @@ class NotificationManagerCompatTest {
                 NotificationManagerCompat.IMPORTANCE_DEFAULT,
             ).build()
 
-        every { tested.getNotificationChannelCompat(testChannel) } returns notificationChannelCompat
-        every { tested.areNotificationsEnabled() } returns false
+        whenever(tested.getNotificationChannelCompat(testChannel))
+            .thenReturn(notificationChannelCompat)
+        whenever(tested.areNotificationsEnabled()).thenReturn(false)
 
         assertFalse(tested.isNotificationChannelEnabled(testChannel))
     }
@@ -100,8 +94,8 @@ class NotificationManagerCompatTest {
     fun `WHEN getNotificationChannelCompat returns null THEN isNotificationChannelEnabled returns false`() {
         val testChannel = "test-channel"
 
-        every { tested.getNotificationChannelCompat(testChannel) } returns null
-        every { tested.areNotificationsEnabled() } returns true
+        whenever(tested.getNotificationChannelCompat(testChannel)).thenReturn(null)
+        whenever(tested.areNotificationsEnabled()).thenReturn(true)
 
         assertFalse(tested.isNotificationChannelEnabled(testChannel))
     }
@@ -112,7 +106,7 @@ class NotificationManagerCompatTest {
 
         ReflectionHelpers.setStaticField(Build.VERSION::class.java, "SDK_INT", 25)
 
-        every { tested.areNotificationsEnabled() } returns true
+        whenever(tested.areNotificationsEnabled()).thenReturn(true)
 
         assertTrue(tested.isNotificationChannelEnabled(testChannel))
     }